apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
sentence-transformers==2.7.0
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
pytest==8.2.1
httpx==0.27.0
ruff==0.4.7
black==24.4.2
